feat: 조직 type에 고등학교 추가, 조직 type별 랭킹을 위한 fragment 추가,
