+++ android/app/src/main/java/com/dragonguard/android/activity/menu/AuthOrgActivity.kt
@@ -64,6 +64,7 @@ class AuthOrgActivity : AppCompatActivity() {
         val arr1 : MutableList<String> = mutableListOf("선택하세요")
         arr1.apply {
             add("대학교")
+            add("고등학교")
             add("회사")
             add("etc")
         }
@@ -113,6 +114,9 @@ class AuthOrgActivity : AppCompatActivity() {
                     }
                     "회사" -> {
                         orgType = "COMPANY"
+                    }
+                    "고등학교" -> {
+                        orgType = "HIGH_SCHOOL"
                     }
                     "etc" -> {
                         orgType = "ETC"
