+++ android/app/src/main/res/layout/activity_search_organization.xml
@@ -90,45 +90,59 @@
                 <View
                     android:layout_width="0dp"
                     android:layout_height="match_parent"
-                    android:layout_weight="0.1"/>
+                    android:layout_weight="0.05"/>
 
-                <RadioGroup
-                    android:id="@+id/org_group"
+                <HorizontalScrollView
                     android:layout_width="0dp"
                     android:layout_height="match_parent"
-                    android:layout_weight="3"
-                    android:gravity="center_vertical"
-                    android:orientation="horizontal">
+                    android:layout_weight="3.3">
+                    <RadioGroup
+                        android:id="@+id/org_group"
+                        android:layout_width="match_parent"
+                        android:layout_height="match_parent"
+                        android:gravity="center_vertical"
+                        android:orientation="horizontal">
 
-                    <RadioButton
-                        android:id="@+id/org_type0"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:fontFamily="@font/ibm_semibold"
-                        android:text="대학"
-                        android:textSize="20sp" />
-                    <RadioButton
-                        android:id="@+id/org_type1"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:fontFamily="@font/ibm_semibold"
-                        android:text="회사"
-                        android:textSize="20sp" />
-                    <RadioButton
-                        android:id="@+id/org_type2"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:fontFamily="@font/ibm_semibold"
-                        android:text="etc"
-                        android:textSize="20sp" />
+                        <RadioButton
+                            android:id="@+id/org_type0"
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"
+                            android:fontFamily="@font/ibm_semibold"
+                            android:text="대학교"
+                            android:textSize="18sp" />
 
-                </RadioGroup>
+                        <RadioButton
+                            android:id="@+id/org_type1"
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"
+                            android:fontFamily="@font/ibm_semibold"
+                            android:text="고등학교"
+                            android:textSize="18sp" />
+                        <RadioButton
+                            android:id="@+id/org_type2"
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"
+                            android:fontFamily="@font/ibm_semibold"
+                            android:text="회사"
+                            android:textSize="18sp" />
+                        <RadioButton
+                            android:id="@+id/org_type3"
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"
+                            android:fontFamily="@font/ibm_semibold"
+                            android:text="etc"
+                            android:textSize="18sp" />
+
+                    </RadioGroup>
+                </HorizontalScrollView>
+
+
 
                 <androidx.appcompat.widget.AppCompatButton
                     android:id="@+id/none_organization"
                     android:layout_width="0dp"
                     android:layout_height="wrap_content"
-                    android:layout_weight="1.6"
+                    android:layout_weight="1.5"
                     android:background="@drawable/round_pink"
                     android:fontFamily="@font/ibm_semibold"
                     android:text="조직 없음"
