+++ android/app/src/main/res/layout/fragment_main.xml
@@ -209,7 +209,7 @@
                             android:layout_width="match_parent"
                             android:layout_height="wrap_content"
                             android:adjustViewBounds="true"
-                            android:layout_marginHorizontal="25dp"
+                            android:layout_marginHorizontal="27dp"
                             android:layout_marginTop="15dp"
                             android:src="@drawable/coins" />
 
@@ -221,7 +221,7 @@
                             android:gravity="top|center_horizontal"
                             android:layout_gravity="bottom"
                             android:textColor="@color/black"
-                            android:textSize="17sp" />
+                            android:textSize="15sp" />
 
                     </LinearLayout>
 
@@ -250,7 +250,7 @@
             <View
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
-                android:layout_weight="0.1" />
+                android:layout_weight="0.05" />
 
 
             <LinearLayout
@@ -323,7 +323,7 @@
                 android:layout_height="0dp"
                 android:layout_gravity="center_horizontal"
                 android:layout_marginHorizontal="20dp"
-                android:layout_weight="0.85"
+                android:layout_weight="0.93"
                 android:orientation="horizontal"
                 android:paddingVertical="5dp">
